+++ meteor/server/migration/migrations.ts
@@ -90,6 +90,9 @@ addSteps1_18_0()
 import { addSteps as addSteps1_19_0 } from './1_19_0'
 addSteps1_19_0()
 
+import { addSteps as addSteps1_32_0 } from './1_32_0'
+addSteps1_32_0()
+
 // Migrations for the in-development release:
 import { addSteps as addStepsX_X_X } from './X_X_X'
 addStepsX_X_X()
